Remove spark version indicator in kernelspecs
Removed spark version (2.1) indicator from the kernelspec directory
names and updated the whitelist example in the start script with new
kernel names.

Also removed some unused or unnecessary variables from start script
and changed culling to the increased recommended value of 12 hours with
note that it should be changed to desired requirements.

Closes #22